requirements.txt
Flask
gunicorn
pyserial
requests